Rename file jsx to js
