+++ app/assets/javascripts/__test__/admin/actors/admin_actor_detail.test.js
@@ -1,7 +1,7 @@
 import React from 'react'
-import AdminActorDetail from './../../../components/admin/actors/admin_actor_detail.jsx'
+import AdminActorDetail from './../../../components/admin/actors/admin_actor_detail.js'
 import { shallow } from 'enzyme'
-jest.unmock('./../../../components/admin/actors/admin_actor_detail.jsx')
+jest.unmock('./../../../components/admin/actors/admin_actor_detail.js')
 
 describe('AdminActorDetailComponent', () => {
   it('管理画面：声優詳細情報が表示されること', () => {
