+++ app/assets/javascripts/__test__/admin/admin.test.js
@@ -1,8 +1,8 @@
 import React from 'react'
 import { shallow } from 'enzyme'
-import Admin from './../../components/admin/admin.jsx'
-import AdminMenu from './../../components/admin/admin_menu.jsx'
-jest.unmock('./../../components/admin/admin.jsx')
+import Admin from './../../components/admin/admin.js'
+import AdminMenu from './../../components/admin/admin_menu.js'
+jest.unmock('./../../components/admin/admin.js')
 
 describe('AnimeComponent', () => {
   it('管理画面が表示されること', () => {
