+++ app/assets/javascripts/__test__/admin/animes/admin_anime.test.js
@@ -1,8 +1,8 @@
 import React from 'react'
 import { shallow } from 'enzyme'
-import AdminAnime from './../../../components/admin/animes/admin_anime.jsx'
-import AdminAnimeDetail from './../../../components/admin/animes/admin_anime_detail.jsx'
-jest.unmock('./../../../components/admin/animes/admin_anime.jsx')
+import AdminAnime from './../../../components/admin/animes/admin_anime.js'
+import AdminAnimeDetail from './../../../components/admin/animes/admin_anime_detail.js'
+jest.unmock('./../../../components/admin/animes/admin_anime.js')
 
 describe('AdminAnimeComponent', () => {
   it('管理画面：アニメ詳細画面が表示されること', () => {
