+++ app/assets/javascripts/__test__/admin/animes/admin_anime_detail.test.js
@@ -1,7 +1,7 @@
 import React from 'react'
-import AdminAnimeDetail from './../../../components/admin/animes/admin_anime_detail.jsx'
+import AdminAnimeDetail from './../../../components/admin/animes/admin_anime_detail.js'
 import { shallow } from 'enzyme'
-jest.unmock('./../../../components/admin/animes/admin_anime_detail.jsx')
+jest.unmock('./../../../components/admin/animes/admin_anime_detail.js')
 
 describe('AdminAnimeDetailComponent', () => {
   it('管理画面：アニメ詳細情報が表示されること', () => {
