+++ app/assets/javascripts/__test__/admin/animes/admin_animes.test.js
@@ -1,8 +1,8 @@
 import React from 'react'
 import { shallow } from 'enzyme'
-import AdminAnimes from './../../../components/admin/animes/admin_animes.jsx'
-import AdminAnimesTable from './../../../components/admin/animes/admin_animes_table.jsx'
-jest.unmock('./../../../components/admin/animes/admin_animes.jsx')
+import AdminAnimes from './../../../components/admin/animes/admin_animes.js'
+import AdminAnimesTable from './../../../components/admin/animes/admin_animes_table.js'
+jest.unmock('./../../../components/admin/animes/admin_animes.js')
 
 describe('AdminAnimesComponent', () => {
   it('管理画面：アニメ情報一覧画面が表示されること', () => {
