+++ app/assets/javascripts/__test__/navbar.test.js
@@ -1,8 +1,8 @@
 import React from 'react'
 import { shallow } from 'enzyme'
 import { Link } from 'react-router'
-import Navbar from './../components/navbar.jsx'
-jest.unmock('./../components/navbar.jsx')
+import Navbar from './../components/navbar.js'
+jest.unmock('./../components/navbar.js')
 
 describe('AnimeComponent', () => {
   it('メニューが表示されること', () => {
