+++ app/assets/javascripts/__test__/welcome/welcome.test.js
@@ -1,8 +1,8 @@
 import React from 'react'
 import { shallow } from 'enzyme'
-import Welcome from './../../components/welcome/welcome.jsx'
-import AnimeList from './../../components/welcome/anime_list.jsx'
-jest.unmock('./../../components/welcome/welcome.jsx')
+import Welcome from './../../components/welcome/welcome.js'
+import AnimeList from './../../components/welcome/anime_list.js'
+jest.unmock('./../../components/welcome/welcome.js')
 
 describe('AnimeComponent', () => {
   it('アニメ一覧画面が表示されること', () => {
