+++ app/assets/javascripts/components.js
@@ -5,4 +5,4 @@
 import React from 'react'
 import ReactDOM from 'react-dom'
 
-import routes from './routes.jsx'
+import routes from './routes.js'
