+++ app/assets/javascripts/components/admin/actors/admin_actor.js
@@ -1,5 +1,5 @@
 import React, { Component } from 'react'
-import AdminActorDetail from './admin_actor_detail.jsx'
+import AdminActorDetail from './admin_actor_detail.js'
 
 export default class AdminActor extends Component {
   render() {
