+++ app/assets/javascripts/components/admin/actors/admin_actor_row.js
