+++ app/assets/javascripts/components/admin/actors/admin_actors.js
@@ -1,5 +1,5 @@
 import React, { Component } from 'react'
-import AdminActorsTable from './admin_actors_table.jsx'
+import AdminActorsTable from './admin_actors_table.js'
 
 export default class AdminActors extends Component {
   render() {
