+++ app/assets/javascripts/components/admin/actors/admin_actors_page.js
