+++ app/assets/javascripts/components/admin/actors/admin_actors_table.js
@@ -1,5 +1,5 @@
-import React, { Component } from 'react'
-import AdminActorRow from './admin_actor_row.jsx'
+import React from 'react'
+import AdminActorRow from './admin_actor_row.js'
 
 export default React.createClass({
   loadActorsFromServer() {
@@ -7,7 +7,7 @@ export default React.createClass({
       url: this.props.url,
       dataType: 'json',
       success: (res) => {
-        this.setState({actor: res})
+        this.setState({actors: res.actors})
       },
       error: (xhr, status, err) => {
         console.error(this.props.url, status, err.toString())
@@ -20,18 +20,18 @@ export default React.createClass({
   },
 
   getInitialState() {
-    return { actor: {} }
+    return { actors: [] }
   },
 
   render() {
     return (
-      <div className='panel panel-default'>
-        <div className='panel-heading'>
-          {this.state.actor.name}
-        </div>
-        <div className='panel-body'>
-        </div>
-      </div>
+      <table className='table'>
+        <tbody>
+          {this.state.actors.map((actor) =>
+            <AdminActorRow actor={actor} key={actor.id} />
+          )}
+        </tbody>
+      </table>
     )
   }
 })
