+++ app/assets/javascripts/components/admin/admin_menu.js
