+++ app/assets/javascripts/components/admin/animes/admin_anime_row.js
