+++ app/assets/javascripts/components/admin/animes/admin_animes_page.js
