+++ app/assets/javascripts/components/app.js
@@ -1,6 +1,6 @@
 import React, { Component } from 'react'
 import { Link } from 'react-router'
-import Navbar from './navbar.jsx'
+import Navbar from './navbar.js'
 
 export default class App extends Component {
   render() {
