+++ app/assets/javascripts/components/navbar.js
