+++ app/assets/javascripts/components/welcome/anime.js
