+++ app/assets/javascripts/components/welcome/anime_list.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import Anime from './anime.jsx';
+import Anime from './anime.js';
 
 export default React.createClass({
   loadAnimesFromServer() {
