+++ app/assets/javascripts/components/welcome/welcome.js
@@ -1,5 +1,5 @@
 import React, { Component } from 'react'
-import AnimeList from './anime_list.jsx'
+import AnimeList from './anime_list.js'
 
 export default class Welcome extends Component {
   render() {
