+++ app/assets/javascripts/index.js
@@ -1,0 +1,4 @@
+import React from 'react'
+import { render } from 'react-dom'
+import App from './components/app.jsx'
+render(<App/>, document.getElementById('app'))
