--- app/views/application/_admin_menu.html.slim
@@ -1,6 +1,0 @@
-ul
-  li= link_to 'TOP', root_path
-  li= link_to '管理TOP', admin_top_path
-  li= link_to 'アニメ', admin_animes_path
-  li= link_to '声優', admin_actors_path
-  li= link_to '音楽', admin_melodies_path
