--- app/views/welcome/index.html.slim
@@ -1,1 +1,0 @@
-= react_component 'AnimeList', url: api_animes_path
